+++ src/ATFlexMessageWindowClass.rs
@@ -4,10 +4,10 @@
 // MVID: F52869E5-0850-48AD-BBBE-68E7A4900AFE
 // Assembly location: C:\Program Files (x86)\Steam\steamapps\common\Shadow Empire\ShadowEmpire.exe
 
-using Microsoft.VisualBasic;
-using Microsoft.VisualBasic.CompilerServices;
-using System;
-using System.Drawing;
+// usingMicrosoft.VisualBasic;
+// usingMicrosoft.VisualBasic.CompilerServices;
+// usingSystem;
+// usingSystem.Drawing;
 
 namespace WindowsApplication1
 {
@@ -59,7 +59,7 @@ namespace WindowsApplication1
           str1 = "";
         }
         SizeF sizeF2 = objgraphics.MeasureString(str2, this.game.VicFont3);
-        DrawMod.DrawTextVic2(ref objgraphics, str2, this.game.VicFont3, (int) Math.Round(400.0 - (double) sizeF2.Width / 2.0), 40 + num1 * 20, this.game.VicColor2, this.game.VicColor2Shade);
+        DrawMod.DrawTextVic2(ref objgraphics, str2, this.game.VicFont3, (int) Math.Round(400.0 -  sizeF2.Width / 2.0), 40 + num1 * 20, this.game.VicColor2, this.game.VicColor2Shade);
       }
       int[] answer = this.Answer;
       let mut tsubpart: SubPartClass =  new TextButtonPartClass(this.game.EditObj.AnswerText[1], 100, tBackbitmap: (ref this.OwnBitmap), bbx: 350, bby: 150);
